+++ Sources/Extensions/SwiftStdlib/ArrayExtensions.swift
@@ -505,6 +505,7 @@ public extension Array where Element: Equatable {
 	///
 	/// - Parameter item: item to check.
 	/// - Returns: an array with all indexes of the given item.
+    @available(*, deprecated: 4.1.1, message: "Use indices(of:) instead", renamed: "indices(of:)")
 	public func indexes(of item: Element) -> [Int] {
 		var indexes: [Int] = []
 		for index in startIndex..<endIndex where self[index] == item {
@@ -512,6 +513,22 @@ public extension Array where Element: Equatable {
 		}
 		return indexes
 	}
+    
+    /// SwifterSwift: All indices of specified item.
+    ///
+    ///        [1, 2, 2, 3, 4, 2, 5].indices(of 2) -> [1, 2, 5]
+    ///        [1.2, 2.3, 4.5, 3.4, 4.5].indices(of 2.3) -> [1]
+    ///        ["h", "e", "l", "l", "o"].indices(of "l") -> [2, 3]
+    ///
+    /// - Parameter item: item to check.
+    /// - Returns: an array with all indices of the given item.
+    public func indices(of item: Element) -> [Int] {
+        var indices: [Int] = []
+        for index in startIndex..<endIndex where self[index] == item {
+            indices.append(index)
+        }
+        return indices
+    }
 	
 	/// SwifterSwift: Remove all instances of an item from array.
 	///