+++ Tests/SwiftStdlibTests/ArrayExtensionsTests.swift
@@ -30,6 +30,11 @@ final class ArrayExtensionsTests: XCTestCase {
 	func testIndexes() {
 		XCTAssertEqual([1, 1, 2, 3, 4, 1, 2, 1].indexes(of: 1), [0, 1, 5, 7])
 	}
+    
+    func testIndices() {
+        XCTAssertEqual([1, 1, 2, 3, 4, 1, 2, 1].indices(of: 1), [0, 1, 5, 7])
+        XCTAssertEqual(["a", "b", "c", "b", "4", "1", "2", "1"].indices(of: "b"), [1, 3])
+    }
 	
 	func testLastIndex() {
 		XCTAssertNotNil([1, 1, 2, 3, 4, 1, 2, 1].lastIndex(of: 2))